Add "table almost full" handling.
Almost full hash tables become very slow. Add detection of and callbacks for
tables which are almost full. A table can be configured to:

* Issue a warning.
* Raise an error.
* Do nothing.
* Resize the table.

Issuing the warning is the default to try to make users aware of the rather
unusual requirement to *make your tables a bit bigger than they need to be*.